clean the mess
