--- app/partials/create-subject.html
@@ -1,106 +1,0 @@
-<form name="postSchema" class="form-horizontal" style="width: 100%" ng-controller="HeaderCtrl as ctrl">
-
-  <!-- Main Card for new subject -->
-  <md-card md-theme-watch="">
-    <md-card-title>
-      <md-card-title-text>
-
-        <span class="md-headline" style="padding-bottom:20px;float:left">New version of a schema under a subject</span>
-
-        <div layout-gt-xs="row" class="layout-gt-xs-row" style="width:100%">
-          <div class="form-group form-inline" style="width: 100%;">
-            <label class="col-sm-1 control-label">Subject </label>
-            <div class="col-sm-5">
-
-              <md-autocomplete
-                required
-                ng-disabled="ctrl.isDisabled"
-                md-no-cache="ctrl.noCache"
-                md-selected-item="ctrl.selectedItem"
-                md-search-text-change="ctrl.searchTextChange(ctrl.searchText)"
-                md-search-text="ctrl.searchText"
-                md-selected-item-change="ctrl.selectedItemChange(item)"
-                md-items="item in ctrl.querySearch(ctrl.searchText)"
-                md-item-text="item.display"
-                md-min-length="1"
-                placeholder="fill in a subject name"
-                style="background:white">
-                <md-item-template>
-                  <span md-highlight-text="ctrl.searchText" md-highlight-flags="^i">{{item.display}}</span>
-                </md-item-template>
-                <div ng-messages="postSchema.autocomplete.$error">
-                  <div ng-message="required">This field is required</div>
-                </div>
-              </md-autocomplete>
-              <!--<input type="email" class="form-control" id="exampleInputEmail1" placeholder="Subject">-->
-            </div>
-
-          </div>
-        </div>
-        <div layout-gt-xs="row" class="layout-gt-xs-row" style="width:100%">
-          <!-- data raw -->
-          <div style="width: 100%" ;>
-            <div ng-hide="curlme" id="newavro"
-                 style="border: 1px solid rgba(0, 0, 0, 0.18);height:330px;border-top-left-radius: 4px;border-top-right-radius: 4px;"
-                 ng-model="newAvroString" ng-readonly="false" ui-ace="{
-                          useWrapMode: true,
-                          showGutter: false,
-                          theme:'chrome',
-                          mode: 'json',
-                          firstLineNumber: 1,
-                          onLoad: aceLoaded,
-                          onChange: aceChanged
-                          }"></div>
-            <div ng-show="curlme" id="curlcommand"
-                 style="border: 1px solid rgba(0, 0, 0, 0.18);height:330px;border-top-left-radius: 4px;border-top-right-radius: 4px;"
-                 ng-model="curlCommand" ng-readonly="true" ui-ace="{
-                          useWrapMode: true,
-                          showGutter: false,
-                          theme:'chrome',
-                          mode: 'batchfile',
-                          firstLineNumber: 1,
-                          onLoad: aceLoaded_,
-                          onChange: aceChanged_
-                          }"></div>
-            <div class="checkbox">
-              <label>
-                <input type="checkbox" ng-model="curlme" ng-checked="isChecked"> Display curl command
-              </label>
-            </div>
-          </div>
-
-        </div>
-      </md-card-title-text>
-    </md-card-title>
-  </md-card>
-
-  <!-- Buttons at the bottom of the form -->
-  <div class="row" style="padding-top:15px;">
-    <center>
-      <button md-ink-ripple
-              class="md-raised btn md-primary md-button md-ink-ripple"
-              ng-click="testCompatibility();"
-              ng-validate="noSubjectName"
-              style="background-color:rgba(56, 87, 119, 0.84)"
-              type="submit"
-              aria-label="Test schema compatibility">
-          <span class="ng-scope">
-          Test compatibility
-          </span>
-        <div class="md-ripple-container"></div>
-      </button>
-      <button md-ink-ripple
-              class="md-raised btn md-primary md-button md-ink-ripple"
-              ng-click="registerNewSchema();"
-              ng-validate="noSubjectName"
-              style="background-color:#49824c"
-              type="submit"
-              aria-label="Register new schema">
-          <span class="ng-scope">
-          Register Schema
-          </span>
-        <div class="md-ripple-container"></div>
-      </button>
-    </center>
-  </div>
-</form>
