--- app/partials/subject.html
@@ -1,109 +1,0 @@
-<div>
-
-  <div class="row" style="padding-bottom:20px;">
-    <div class="btn-toolbar" role="toolbar" style="margin-left: 5px;">
-
-      <!-- 1. subject-name -->
-      <div style="padding-top: 5px;margin-left: 10px;margin-right: 10px;float: left;min-width:250px;">
-        <b>Subject:</b> {{subjectObject.subjectName}}
-      </div>
-
-      <!-- 2. Avro / Table -->
-      <div ng-show="tableViewOn">
-        <div class="btn-group" role="group">
-          <button type="button" class="btn btn-sm btn-default" ng-click="changeView()"><span
-            class="glyphicon"></span> Avro
-          </button>
-          <button type="button" class="btn btn-sm btn-default active" ng-click="changeView()">
-            <span class="glyphicon glyphicon-th-list"></span> Table
-          </button>
-        </div>
-      </div>
-      <div ng-hide="tableViewOn">
-        <div class="btn-group" role="group">
-          <button type="button" class="btn btn-sm btn-default active" ng-click="changeView()">
-            <span class="glyphicon"></span> Avro
-          </button>
-          <button type="button" class="btn btn-sm btn-default" ng-click="changeView()"><span
-            class="glyphicon glyphicon-th-list"></span> Table
-          </button>
-        </div>
-      </div>
-
-      <!-- 3. version -->
-      <div style="padding-top: 5px;margin-left: 10px;margin-right: 10px;float: left;">
-        version
-      </div>
-      <div ng-show="multipleVersionsOn" class="btn-group">
-        <button class="btn btn-sm btn-success dropdown-toggle" type="button" data-toggle="dropdown"
-                aria-haspopup="true" aria-expanded="false">
-          {{subjectObject.version}} <span class="caret"></span>
-        </button>
-        <ul class="dropdown-menu">
-          <li><a ng-repeat="v in subjectObject.otherVersions | orderBy:'v'"
-                 ng-href="/#/subject/{{subjectObject.subjectName}}/version/{{v}}">
-            {{v}} </a></li>
-        </ul>
-      </div>
-      <div ng-hide="multipleVersionsOn" class="btn-group" style="margin-left: 0;padding-top: 3px;">
-              <span class="label label-default"
-                    style="background-color:#286090">{{subjectObject.version}}</span>
-      </div>
-
-      <!-- 4 schema id -->
-      <div style="padding-top: 5px;margin-left: 10px;margin-right: 10px;float: left;">
-        schema id
-      </div>
-      <div class="btn-group" style="margin-left: 0;padding-top: 3px;">
-        <span class="label label-default" style="background-color:#286090">{{subjectObject.id}}</span>
-      </div>
-
-    </div>
-
-  </div>
-
-  <!-- 3a. data table -->
-  <div ng-show="tableViewOn">
-    <div class="panel panel-default">
-      <div class="panel-body">
-        <p style="padding-top: 10px">
-          <b>type:</b> {{subjectObject.Schema.type}} |
-          <b>name:</b> {{subjectObject.Schema.name}} |
-          <b>namespace:</b> {{subjectObject.Schema.namespace}}</p>
-      </div>
-    </div>
-
-    <table class="table table-bordered table-striped table-hover">
-      <thead>
-      <tr>
-        <th width="30%">Name</th>
-        <th width="40%">Type</th>
-        <th width="20%">Default</th>
-      </tr>
-      </thead>
-      <tbody>
-      <tr ng-repeat="i in subjectObject.Schema.fields">
-        <td>{{i.name}}</td>
-        <td>{{i.type}}</td>
-        <td>{{i.default}}</td>
-      </tr>
-      </tbody>
-    </table>
-  </div>
-
-  <!-- 3a. data raw -->
-  <div ng-hide="tableViewOn">
-    <div id="left"
-         style="border: 1px solid rgba(0, 0, 0, 0.18);border-top-left-radius: 4px;border-top-right-radius: 4px;"
-         ng-model="aceString" ng-readonly="true" ui-ace="{
-                          useWrapMode: true,
-                          showGutter: false,
-                          theme:'chrome',
-                          mode: 'json',
-                          firstLineNumber: 1,
-                          onLoad: aceLoaded,
-                          onChange: aceChanged
-                          }"></div>
-  </div>
-
-</div> <!-- end if selected subject -->
