+++ app/src/home/home.controller.js
@@ -1,0 +1,4 @@
+schemaRegistryUIApp.controller('HomeCtrl', function ($scope, $routeParams, $mdToast, $log, schemaRegistryFactory) {
+  $mdToast.hide();
+  schemaRegistryFactory.visibleCreateSubjectButton(true);
+});
