+++ app/src/home/home.html
@@ -1,7 +1,7 @@
 <div class="container-fluid-centered" style="height: 500px;">
   <div class="row-fluid">
     <div class="centering text-center">
-      <p> Welcome to the <b>schema-registry-ui</b> v0.3</p><br/>
+      <p> Welcome to the <b>schema registry ui</b></p><br/>
       <p> select a schema or <a ng-href="#/create-subject/">create a new one</a></p><br/>
     </div>
   </div>
