+++ bower.json
@@ -25,7 +25,8 @@
     "angular-material": "^1.0.9",
     "angular-animate": "^1.5.7",
     "angular-aria": "^1.5.7",
-    "angular-route": "^1.5.7"
+    "angular-route": "^1.5.7",
+    "font-awesome": "fontawesome#^4.6.3"
   },
   "resolutions": {
     "angular": "~1.4.*"
